Change shell parsing to match shortest non-ambiguous command
Previously shell commands were parsed by checking the first character
only. Changed the parser to match the shortest non-ambiguous command,
in such a way that invalid commands are not parsed as the valid command
that starts with the same character, and current command shortcuts are
retained as long as they are non-ambiguous.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>