Change the kernel file management
Instead of symlinkin to each of kernel/*.c file, create a master symlink
to $hdadir -> hda directory, and include each file from "../hda/xxx.c"
in kernel/*.c file.  This makes the kernel/ subdirectory more consistent.

The configure script is modified for that, and also for defining the
conditional for each patch_*.c so that kernel/Makefile.am can compile
files selectively.

Signed-off-by: Takashi Iwai <tiwai@suse.de>