Support HD-audio core library for kernel v4.1+
Now the HD-audio code is split and the core helper codes are put in
sound/hda/ directory.  hda-emu also follows this change.

The --with-hdadir configure option is replaced with --with-kerneldir
option.  Instead of passing the HD-audio directory, pass the root
directory of Linux kernel here.

configure script has been a bit modified to take care of two
directories, sound/hda and sound/pci/hda.  The former is built in
lib subdirectory while the latter is still symlinked to hda
subdirectory and built in hda subdirectory.